@@ -10,11 +10,12 @@
     "serve": "vitepress serve content"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^1.2.2",
+    "@vitejs/plugin-vue": "^1.9.3",
     "gray-matter": "^4.0.3",
     "sass": "^1.32.12",
-    "vitepress": "^0.15.4",
-    "vue": "^3.0.11"
+    "typescript": "^4.4.4",
+    "vitepress": "^0.20.0",
+    "vue": "^3.2.20"
   },
   "engines": {
     "node": ">=12.0.0"
